requirements.txt
fastapi
pydantic==2.7.0
requests==2.31.0
python-dotenv
uvicorn